requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
openpyxl==3.0.10
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
